Fix critical SEO bug: Move structured data to server component
ISSUE:
Structured data (JSON-LD scripts) were inside PricingContent, a client component marked with 'use client'. Client components hydrate after the initial HTML is sent, so search engine crawlers never saw the schema markup, defeating the entire purpose of SEO optimization.

FIX:
- Split pricing page into server and client components
- Created PricingContent.tsx as separate client component file with 'use client' directive
- Moved page.tsx to be a server component (no 'use client')
- Generate all structured data at build time in server component
- Render Script tags with JSON-LD in server component (outside Suspense boundary)
- Use strategy="beforeInteractive" to ensure schemas load early

RESULT:
- Structured data now in initial HTML sent to crawlers ✓
- Search engines can properly index FAQPage, Breadcrumb, and SoftwareApplication schemas ✓
- Zero functional changes to user experience ✓
- Build passes successfully ✓

Files Changed:
- app/pricing/page.tsx: Now a server component that generates schemas
- app/pricing/PricingContent.tsx: New file with all client-side logic

This is a critical fix for SEO functionality. Without it, none of the structured data optimizations would work.
